migrate ReplyList and ReplyCard to TypeScript
